Generalize BlockFilterFunc
It would be great if we could treat ResourceFilter and other block
filters together in an orthogonal way and, at the same time, provide
derived types such as Resource to library users. To do this, We start
using Generics here to avoid code duplication, but it is not runtime
safe because of the interface cast. We reached this design decision
after balancing the convenience of library users and maintainers.
However, since we are not confident that the design is correct, we will
only expose the derived type aliases, and the generic implementation is
kept private.