Restrict blocks to which VerticalFormat is applied
When the block type checks were transferred to the respective rewriting
rules, the vertical format was unintentionally applied to all block
types, resulting in differing test results.  While there is another
problem of vertical format not supporting one-line blocks, we would like
to limit the scope of informal vertical format as much as possible.
Therefore, we made it possible to restrict the block type of vertical
format.