Rename references for website_domain and website_endpoint
In refactoring, some attributes need to rewrite their references to a
new resource. To implement this, a simple string substitution is
insufficient and requires an understanding of HCL expressions.

At first, my intuition was to call the RenameVariablePrefix function in
hclwrite, which can only be used with forward matching. However, in our
use case, the resource name is arbitrary and cannot be determined by
resource type alone. The only way is to enumerate all variable
references in the expression and filter the target.

Therefore, I implemented it by recursively inspecting all attributes of
all blocks, including nested blocks, to create a list of variable
references and then using regular expressions to build arguments and
replace them. As you know, this is a bit inefficient, but it is the only
way to implement it using the current public API of hclwrite.

Variable in HCL terminology can be confused with input variable in the
context of Terraform, so I have chosen the name Reference. Strictly
speaking, there is also the term "Named Value" in the Terraform
documentation, but it does not seem widely used. If the term Reference
is ambiguous, it may change in the future.