Mock iam_endpoint explicitly for testing with localstack
Starting from Terraform v1.6, the s3 backend configuration has changed,
and this caused the acceptance test to fail. We can workaround the issue
by specifying the `iam_endpoint`. The `iam_endpoint` argument has been
deprecated since Terraformv1.6. It is recommended to use the newly added
`endpoints` argument, though, for running tests on multiple Terraform
versions, the test settings must remain compatible with the older
versions. This is a test-only issue, so I'll use the deprecated
`iam_endpoint` for now.

FYI:
https://github.com/minamijoyo/tfmigrate/pull/157/commits/7591205b3f296e07e903a10d5347467a4ad0b868